fix: color vasr and px to em
