+++ src/pages/ProtectedPages/Student/Student.styles.scss
@@ -1,0 +1,5 @@
+.register-button {
+  background-color: var(--color-white);
+  padding: 16px 8px;
+  box-shadow: 4px 4px 12px var(--text-shadow-color);
+}
